+++ src/Pages/Validator/Components/index.js
@@ -1,6 +1,7 @@
-export { default as ValidatorSpotlight } from './ValidatorSpotlight';
 export { default as ValidatorCommission } from './ValidatorCommission';
 export { default as ValidatorDelegations } from './ValidatorDelegations';
-export { default as ValidatorUnbondingDelegations } from './ValidatorUnbondingDelegations';
 export { default as ValidatorDelegationTxs } from './ValidatorDelegationTxs';
+export { default as ValidatorGovernance } from './ValidatorGovernance';
+export { default as ValidatorSpotlight } from './ValidatorSpotlight';
+export { default as ValidatorUnbondingDelegations } from './ValidatorUnbondingDelegations';
 export { default as ValidationTxs } from './ValidationTxs';